[Build] Fix build failed error

+++ seatunnel-connectors-v2/connector-dolphindb-wt/src/test/java/org/apache/seatunnel/connectors/dolphindb/utils/DolphinDBSaveModeUtilTest.java
@@ -18,11 +18,12 @@ public class DolphinDBSaveModeUtilTest {
     void fillingCreateSql() {
         List<Column> columns = new ArrayList<>();
 
-        columns.add(PhysicalColumn.of("id", BasicType.LONG_TYPE, null, true, null, ""));
-        columns.add(PhysicalColumn.of("name", BasicType.STRING_TYPE, null, true, null, ""));
-        columns.add(PhysicalColumn.of("age", BasicType.INT_TYPE, null, true, null, ""));
-        columns.add(PhysicalColumn.of("gender", BasicType.BYTE_TYPE, null, true, null, ""));
-        columns.add(PhysicalColumn.of("create_time", BasicType.LONG_TYPE, null, true, null, ""));
+        columns.add(PhysicalColumn.of("id", BasicType.LONG_TYPE, (Long) null, true, null, ""));
+        columns.add(PhysicalColumn.of("name", BasicType.STRING_TYPE, (Long) null, true, null, ""));
+        columns.add(PhysicalColumn.of("age", BasicType.INT_TYPE, (Long) null, true, null, ""));
+        columns.add(PhysicalColumn.of("gender", BasicType.BYTE_TYPE, (Long) null, true, null, ""));
+        columns.add(
+                PhysicalColumn.of("create_time", BasicType.LONG_TYPE, (Long) null, true, null, ""));
 
         String result =
                 DolphinDBSaveModeUtil.fillingCreateSql(
